move theme toggle
